updated list after retry
