+++ learn/src/main/java/org/navgurukul/learn/ui/learn/AssessmentFragment.kt
@@ -123,7 +123,6 @@ class AssessmentFragment : Fragment() {
             }
         }
 
-
     private fun setupIncorrectOutputLayout(list: List<BaseCourseContent>) {
         mBinding.incorrectOutputLayout.btnSeeExplanation.setOnClickListener {
             initIncorrectRV(list)
@@ -133,21 +132,10 @@ class AssessmentFragment : Fragment() {
         }
 
         mBinding.incorrectOutputLayout.btnRetry.setOnClickListener {
-            contentAdapter.submitList(resetList())
+            fragmentViewModel.handle(AssessmentFragmentViewModel.AssessmentFragmentViewActions.ShowUpdatedOutput)
             mBinding.incorrectOutputLayout.isVisible = false
             isContentRvClickable = true
         }
-    }
-
-    private fun resetList(): MutableList<BaseCourseContent>? {
-        val newList = fragmentViewModel.viewState.value?.assessmentContentListForUI?.toMutableList()
-        newList?.forEach {
-            if(it.component == BaseCourseContent.COMPONENT_OPTIONS){
-                    val item = it as OptionsBaseCourseContent
-                    item.value = item.value.toMutableList().map{ it.copy(viewState = OptionViewState.NOT_SELECTED) }
-            }
-        }
-        return newList
     }
 
     private fun getNewReferencedList(list: List<BaseCourseContent>?): List<BaseCourseContent>? {
@@ -158,19 +146,6 @@ class AssessmentFragment : Fragment() {
                 )
             }else{
                 it
-            }
-        }
-        return newList
-    }
-
-    private fun getSingleSelectedOptionInNewList(selectedOption: OptionResponse): MutableList<BaseCourseContent>? {
-        val newList = fragmentViewModel.viewState.value?.assessmentContentListForUI?.toMutableList()
-        newList?.forEach {
-            if(it.component == BaseCourseContent.COMPONENT_OPTIONS){
-                    val item = it as OptionsBaseCourseContent
-                    item.value = item.value.toMutableList().map{ item ->
-                        item.copy(viewState = if(item == selectedOption) OptionViewState.SELECTED else OptionViewState.NOT_SELECTED)
-                    }
             }
         }
         return newList
